+++ hc/api/views.py
@@ -12,7 +12,7 @@ from uuid import UUID
 from cronsim import CronSim, CronSimError
 from django.conf import settings
 from django.core.signing import BadSignature
-from django.db import connection
+from django.db import connection, transaction
 from django.db.models import Prefetch
 from django.http import (
     Http404,
@@ -449,11 +449,19 @@ def update_check(request: ApiRequest, code: UUID) -> HttpResponse:
 
 @authorize
 def delete_check(request: ApiRequest, code: UUID) -> HttpResponse:
+    # Don't acquire lock right away, first see if the check exists
+    # and matches the API key
     check = get_object_or_404(Check, code=code)
     if check.project_id != request.project.id:
         return HttpResponseForbidden()
 
-    check.lock_and_delete()
+    # Start a transaction, select for update, delete.
+    # Use get_object_or_404 here again, in case another concurrent request
+    # has *just* deleted this check.
+    with transaction.atomic():
+        check = get_object_or_404(Check.objects.select_for_update(), code=code)
+        check.delete()
+
     return JsonResponse(check.to_dict(v=request.v))